feat(v2): Implement generic account serializer API
